Updating API endpoints
Refactoring and cleanup
