+++ src/app/api/collections/vocabs.py
@@ -1,0 +1,30 @@
+from __future__ import annotations
+
+from uuid import UUID
+
+from aiohttp import ClientSession
+
+from app.api.collections.models import CollectionTreeNode
+from app.core.constants import COLLECTION_ROOT_ID
+
+
+async def tree_from_vocabs(session: ClientSession, node_id: UUID):
+    url = f"https://vocabs.openeduhub.de/w3id.org/openeduhub/vocabs/oehTopics/{node_id}.json"
+    response = await session.get(url=url)
+    if response.status == 200:
+        data = await response.json()
+        keyword = "hasTopConcept" if str(node_id) == COLLECTION_ROOT_ID else "narrower"
+        return collection_to_model(data[keyword])
+
+
+def collection_to_model(data: list[dict]) -> list[CollectionTreeNode]:
+    return [
+        CollectionTreeNode(
+            noderef_id=collection["id"].split("/")[-1],
+            title=collection["prefLabel"]["de"],
+            children=collection_to_model(collection["narrower"])
+            if "narrower" in collection
+            else [],
+        )
+        for collection in data
+    ]
